Smaller dtype

+++ ScaFFold/datagen/volumegen.py
@@ -179,8 +179,9 @@ def main(config: Dict):
                 0,
                 dtype=np.float32,
             )
+            # This contains values 0 <= x <= n_categories
             mask = np.full(
-                (config.vol_size, config.vol_size, config.vol_size), 0, dtype=np.int64
+                (config.vol_size, config.vol_size, config.vol_size), 0, dtype=np.uint8
             )
 
             global_vol_idx = curr_vol[0]
@@ -226,7 +227,8 @@ def main(config: Dict):
             volume_to_save = np.ascontiguousarray(
                 volume.transpose((3, 0, 1, 2)), dtype=np.float32
             )
-            mask_to_save = np.ascontiguousarray(mask, dtype=np.int64)
+            # This contains values 0 <= x <= n_categories
+            mask_to_save = np.ascontiguousarray(mask, dtype=np.uint8)
 
             vol_file = os.path.join(vol_path, subdir, f"{global_vol_idx}.npy")
             with open(vol_file, "wb") as f:
